Renamed the component for dashboard to home because it makes more sense.
